+++ src/ctf01d/employees/employ_web_server.cpp
@@ -332,14 +332,7 @@ void employ_web_server::updateJsonCache() {
 
   for (unsigned int i = 0; i < m_config->teams().size(); i++) {
       ctf01d::team_config teamConf = m_config->teams()[i];
-      nlohmann::json teamInfo;
-      teamInfo["id"] = teamConf.id();
-      teamInfo["name"] = teamConf.name();
-      teamInfo["ip_address"] = teamConf.ip_or_host();
-      teamInfo["logo"] = "./logo/team/" + teamConf.id();
-      teamInfo["logo-big"] = "./logo/big/team/" + teamConf.id();
-      teamInfo["logo_last_write_time"] = teamConf.get_logo_last_modified_time();
-
+      nlohmann::json teamInfo = teamConf.to_json();
       jsonGame["teams"].push_back(teamInfo);
       jsonTeams["teams"].push_back(teamInfo);
   }