+++ src/ctf01d/include/ctf01d_globals.h
@@ -49,6 +49,13 @@ class json_fields {
 public:
   inline static const std::string SUMMARY_ACTIVITIES = "sum_act";
   inline static const std::string TRIES = "tries";
+  inline static const std::string ID = "id";
+  inline static const std::string NAME = "name";
+  inline static const std::string DESCRIPTION = "dsc";
+  inline static const std::string UPDATED = "upd";
+  inline static const std::string IP_OR_HOST = "ip_address";
+  inline static const std::string LOGO = "logo";
+  inline static const std::string LOGO_BIG = "logo-big";
 };
 
 class yaml_keys {
@@ -61,6 +68,17 @@ public:
   inline static const std::string COFFEE_BREAK = "coffee-break";
   inline static const std::string FLAG_LIFETIME_IN_SECONDS = "flag-lifetime-in-seconds";
   inline static const std::string FLAG_COST_IN_POINTS = "flag-cost-in-points";
+  inline static const std::string TYPE = "type";
+  inline static const std::string TYPE_NORMAL = "normal";
+  inline static const std::string TYPE_RED = "red";
+  inline static const std::string TYPE_BLUE = "blue";
+  inline static const std::string TYPE_QUEST = "guest";
+  inline static const std::string TYPE_DISQUALIFIED = "disqualified";
+  inline static const std::string DESCRIPTION = "description";
+  inline static const std::string ACTIVE = "active";
+  inline static const std::string LOGO = "logo";
+  inline static const std::string LOGO_BIG = "logo-big";
+  inline static const std::string IP_OR_HOST = "ip-or-host";
 };
 
 } // namespace ctf01d