+++ src/wsjcpp_resources/__data_sample_html_css_scoreboard_css_path67c217.cpp
@@ -17,13 +17,13 @@ const std::string &RES___data_sample_html_css_scoreboard_css_path67c217::getPack
 // ---------------------------------------------------------------------
 
 int RES___data_sample_html_css_scoreboard_css_path67c217::getBufferSize() const {
-  return 21376;
+  return 21509;
 }
 
 // ---------------------------------------------------------------------
 
 const char *RES___data_sample_html_css_scoreboard_css_path67c217::getBuffer() const {
-  static const std::string sRet = "" // size: 21376
+  static const std::string sRet = "" // size: 21509
     "@font-face {\n"
     "  font-family: 'DejaVuSans';\n"
     "  src: url('../font/AmazingGroteskDemi.otf');\n"
@@ -139,6 +139,14 @@ const char *RES___data_sample_html_css_scoreboard_css_path67c217::getBuffer() co
     "\n"
     ".team-name {\n"
     "  font-size: 20px;\n"
+    "  white-space: nowrap;\n"
+    "  overflow: hidden;\n"
+    "  text-overflow: ellipsis;\n"
+    "}\n"
+    "\n"
+    ".team-description {\n"
+    "  font-size: 10px;\n"
+    "  line-height: 10px;\n"
     "  white-space: nowrap;\n"
     "  overflow: hidden;\n"
     "  text-overflow: ellipsis;\n"