+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Renamed 'flag_lifetime_in_seconds' -> 'flag-lifetime-in-seconds'
 - Renamed 'flag_cost_in_points' -> 'flag-cost-in-points'
 - Redesign 'coffee-break' section
+- Fixed #121. Add team description (like a town or university) and show it in scoreboard small text under team name
 
 ## [v0.7.3] - (2026 June 10)